+++ mayan/apps/document_states/html_widgets.py
@@ -1,6 +1,8 @@
 from django.template.loader import render_to_string
 from django.utils.html import format_html_join
 
+from mayan.apps.navigation.html_widgets import SourceColumnWidget
+
 
 def widget_transition_events(transition):
     return format_html_join(
@@ -12,12 +14,5 @@ def widget_transition_events(transition):
     )
 
 
-class WorkflowLogExtraDataWidget:
+class WorkflowLogExtraDataWidget(SourceColumnWidget):
     template_name = 'document_states/extra_data.html'
-
-    def render(self, name=None, value=None):
-        return render_to_string(
-            template_name=self.template_name, context={
-                'value': value
-            }
-        )